Updating

+++ CubitOpenSource/Estoque/scripts/product/delete.php
@@ -1,5 +1,10 @@
 <?php
-if (! empty($id)) {
-	$this->dbAdmin->findTable("products")->delete($id);
+if (! empty($ids)) {
+	$ids = explode("-", $ids);
+	foreach ($ids as $id) {
+		if (! empty($id)) {
+			$this->dbAdmin->findTable("products")->delete($id);
+		}
+	}	
 }
 $this->util->redirect("products");
